+++ qbft/spectest/tests/timeout/timeoutduration/round_3_duration.go
@@ -13,7 +13,7 @@ func Round3Duration() *tests.MultiSpecTest {
 	testingNetwork := types.HoleskyNetwork
 	height := qbft.FirstHeight
 	var round qbft.Round = 3
-	dutyStartTime := uint64(testingNetwork.EstimatedTimeAtSlot(phase0.Slot(height)))
+	dutyStartTime := testingNetwork.EstimatedTimeAtSlot(phase0.Slot(height))
 
 	return &tests.MultiSpecTest{
 		Name: "round 3 duration",
